Update Decimal -> Double

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.pub/src/main/java/nts/uk/ctx/at/record/pub/remainnumber/annualleave/AnnLeaveOfThisMonth.java
@@ -23,7 +23,7 @@ public class AnnLeaveOfThisMonth {
 	/**
 	 * 付与日数
 	 */
-	private BigDecimal grantDays;
+	private Double grantDays;
 	
 	/**
 	 * 月初残日数
